Struct jet_bonds::control::instructions::initialize_bond_manager::InitializeBondManager

pub struct InitializeBondManager<'info> {

    pub bond_manager: AccountLoader<'info, BondManager>,
    pub underlying_token_vault: Box<Account<'info, TokenAccount>>,
    pub underlying_token_mint: Box<Account<'info, Mint>>,
    pub bond_ticket_mint: Box<Account<'info, Mint>>,
    pub claims: Box<Account<'info, Mint>>,
    pub collateral: Box<Account<'info, Mint>>,
    pub authority: Signer<'info>,
    pub airspace: AccountInfo<'info>,
    pub underlying_oracle: AccountInfo<'info>,
    pub ticket_oracle: AccountInfo<'info>,
    pub fee_destination: Box<Account<'info, TokenAccount>>,
    pub payer: Signer<'info>,
    pub rent: Sysvar<'info, Rent>,
    pub token_program: Program<'info, Token>,
    pub system_program: Program<'info, System>,

}

Initialize a BondManager The BondManager acts as a sort of market header. Responsible for coordination and authorization of the accounts utilized and interacting with the program

Fields
bond_manager: AccountLoader<'info, BondManager>

The BondManager manages asset tokens for a particular bond duration

underlying_token_vault: Box<Account<'info, TokenAccount>>

The vault for storing the token underlying the bond tickets

underlying_token_mint: Box<Account<'info, Mint>>

The mint for the assets underlying the bond tickets

bond_ticket_mint: Box<Account<'info, Mint>>

The minting account for the bond tickets

claims: Box<Account<'info, Mint>>

Mints tokens to a margin account to represent debt that must be collateralized

collateral: Box<Account<'info, Mint>>

Mints tokens to a margin account to represent debt that must be collateralized

authority: Signer<'info>

The authority that must sign to make this change

airspace: AccountInfo<'info>

The airspace being modified

underlying_oracle: AccountInfo<'info>

The oracle for the underlying asset price CHECK: determined by caller

ticket_oracle: AccountInfo<'info>

The oracle for the bond ticket price CHECK: determined by caller

fee_destination: Box<Account<'info, TokenAccount>>

The account where fees are allowed to be withdrawn

payer: Signer<'info>

The account paying rent for PDA initialization

rent: Sysvar<'info, Rent>

Rent sysvar

token_program: Program<'info, Token>

SPL token program

system_program: Program<'info, System>

Solana system program