Struct jet_bonds::control::state::BondManager

#[repr(C)]
pub struct BondManager {

    pub version_tag: u64,
    pub airspace: Pubkey,
    pub orderbook_market_state: Pubkey,
    pub event_queue: Pubkey,
    pub asks: Pubkey,
    pub bids: Pubkey,
    pub underlying_token_mint: Pubkey,
    pub underlying_token_vault: Pubkey,
    pub bond_ticket_mint: Pubkey,
    pub claims_mint: Pubkey,
    pub collateral_mint: Pubkey,
    pub underlying_oracle: Pubkey,
    pub ticket_oracle: Pubkey,
    pub seed: [u8; 32],
    pub orderbook_paused: bool,
    pub tickets_paused: bool,
    pub duration: i64,
    pub nonce: u64,
    /* private fields */

}

The BondManager contains all the information necessary to run the bond market

Utilized by program instructions to verify given transaction accounts are correct. Contains data about the bond market including the tenor and ticket<->token conversion rate

Fields
version_tag: u64

Versioning and tag information

airspace: Pubkey

The airspace the market is a part of

orderbook_market_state: Pubkey

The market state of the agnostic orderbook

event_queue: Pubkey

The orderbook event queue

asks: Pubkey

The orderbook asks byteslab

bids: Pubkey

The orderbook bids byteslab

underlying_token_mint: Pubkey

The token mint for the underlying asset of the bond tickets

underlying_token_vault: Pubkey

Token account storing the underlying asset accounted for by this ticket program

bond_ticket_mint: Pubkey

The token mint for the bond tickets

claims_mint: Pubkey

Mint owned by bonds to issue claims against a user. These claim notes are monitored by margin to ensure claims are repaid.

collateral_mint: Pubkey

Mint owned by bonds to issue collateral value to a user The collateral notes are monitored by the margin program to track value

underlying_oracle: Pubkey

oracle that defines the value of the underlying asset

ticket_oracle: Pubkey

oracle that defines the value of the bond tickets

seed: [u8; 32]

The user-defined part of the seed that generated this bond manager’s PDA

orderbook_paused: bool

Is the market taking orders

tickets_paused: bool

Can tickets be redeemed

duration: i64

Units added to the initial stake timestamp to determine claim maturity

nonce: u64

Used to generate unique order tags

Implementations

impl BondManager

pub fn authority_seeds(&self) -> [&[u8]; 5]

for signing CPIs with the bond manager account