Module jet_bonds::events

Structs

BondManagerInitialized

EventAdapterRegistered

MarginUserInitialized

ObligationCreated

ObligationFulfilled

ObligationRepay

OrderCancelled

OrderPlaced

OrderbookInitialized

PositionRefreshed

SkippedError

Error that was ignored because it can be handled, and it is more important to complete the instruction.

TicketRedeemed

TicketTransferred

TicketsStaked

ToggleOrderMatching

TokensExchanged

Enums

OrderType