Module jet_bonds::instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Modules

state

Instruction struct definitions for #[state] methods.

Structs

AuthorizeCrank

Instruction.

CancelOrder

Instruction.

ConsumeEvents

Instruction.

ExchangeTokens

Instruction.

InitializeBondManager

Instruction.

InitializeMarginUser

Instruction.

InitializeOrderbook

Instruction.

LendOrder

Instruction.

MarginBorrowOrder

Instruction.

ModifyBondManager

Instruction.

PauseOrderMatching

Instruction.

PopAdapterEvents

Instruction.

RedeemTicket

Instruction.

RefreshPosition

Instruction.

RegisterAdapter

Instruction.

Repay

Instruction.

ResumeOrderMatching

Instruction.

RevokeCrank

Instruction.

SellTicketsOrder

Instruction.

Settle

Instruction.

StakeBondTickets

Instruction.

TranferTicketOwnership

Instruction.