Module jet_bonds::orderbook::state

State utilities and structs for the Orderbook

Structs

CallbackFlags

Binary flags for the CallbackInfo

CallbackInfo

The CallbackInfo is information about an order that is stored in the Event Queue used to manage order metadata

EventAdapterMetadata

EventQueue

todo: algebraic type parameter could provide compile time checks on use as market or user adapter

FillInfo

OrderParams

Parameters needed for order placement

OrderTag

OrderbookMut

Set of accounts that are commonly needed together whenever the orderbook is modified

OutInfo

QueueIterator

Utility struct for iterating over a queue

Enums

OrderbookEvent

Constants

TICK_SIZE

The tick_size used in fp32 operations on the orderbook

Traits

Event

WithQuoteQty

Trait to retrieve the posted quote values from an OrderSummary

Functions

event_queue_len

Calculated the length of the event queue buffer, given the maximum number of events

fp32_div

a is fp0, b is fp32 and result is a/b fp0

fp32_mul

Multiply a u64 with a fixed point 32 number a is fp0, b is fp32 and result is a*b fp0

orderbook_slab_len

Find the len of the byteslab representing an orderbook side, given the maximum number of orders