Struct jet_bonds::orderbook::state::CallbackInfo

#[repr(C)]
pub struct CallbackInfo {
    pub order_tag: OrderTag,
    pub owner: Pubkey,
    pub fill_account: Pubkey,
    pub out_account: Pubkey,
    pub adapter_account_key: Pubkey,
    pub order_submitted: [u8; 8],
    pub flags: CallbackFlags,
    /* private fields */
}

The CallbackInfo is information about an order that is stored in the Event Queue used to manage order metadata

Fields
order_tag: OrderTag

The order tag is generated by the program when submitting orders to the book Used to seed and track PDAs such as Obligation

owner: Pubkey

authority who signed to place the order and is permitted to cancel the order. for margin orders, this is the owner of the MarginUser, which is the margin account PDA from the margin program.

fill_account: Pubkey

the account that will be assigned ownership of any output resulting from a fill. for margin orders this is the margin user. for auto-stake, this account will be set as the split ticket owner. otherwise this is the token account to be deposited into.

out_account: Pubkey

margin user or token account to be deposited into on out the account that will be assigned ownership of any output resulting from an out. for margin orders this is the margin user. otherwise this is the token account to be deposited into.

adapter_account_key: Pubkey

Pubkey of the account that will recieve the event information

order_submitted: [u8; 8]

The unix timestamp for the slot that the order entered the aaob

flags: CallbackFlags

configuration used by callback execution

Implementations

impl CallbackInfo

pub const LEN: usize = 167usize

pub fn new(
bond_manager_key: Pubkey,
owner: Pubkey,
fill_account: Pubkey,
out_account: Pubkey,
adapter: Pubkey,
order_submitted: UnixTimestamp,
flags: CallbackFlags,
nonce: u64
) -> Self

pub fn adapter(&self) -> Option<Pubkey>

pub fn order_submitted_timestamp(&self) -> UnixTimestamp