Module jet_bonds::tickets

Program instructions and structs related to the redeemable bond tickets

Bond Tickets

The program abstracts the concept of a bond into bond tickets. Bond tickets are fungible spl tokens that must be staked to claim their underlying value. In order to create bond tickets, a user must either place a lend order on the orderbook, or exchange the token underlying the bond market (in practice, almost never will users do this, as it locks their tokens for at least the tenor of the market).

Ticket kinds

The program allots for two types of ticket redemption. The ClaimTicket is given when a user stakes directly with the program. As there is no information about the creation of the tickets, a ClaimTicket does not have accounting for principal or interest, and only contains a redemptive value.

Conversely, a SplitTicket contains split principal and interest values. As well as the slot it was minted. To create a SplitTicket, you must configure your OrderParams auto_stake flag to true. This will allow to program to immediately stake your tickets as the match event is processed.

Modules

instructions

Program instructions for using bond tickets

state

Methods and structs for defining bond tickets