Struct jet_bonds::tickets::state::SplitTicket

pub struct SplitTicket {
    pub owner: Pubkey,
    pub bond_manager: Pubkey,
    pub order_tag: OrderTag,
    pub struck_timestamp: i64,
    pub maturation_timestamp: i64,
    pub principal: u64,
    pub interest: u64,
}

A split ticket represents a claim of underlying tokens as the result of a lending action.

The split ticket is generated when a user places a matched order with the auto_stake flag set to true. By taking the difference between the matched base and quote quantities, the split ticket assigns principal and interest values.

Fields
owner: Pubkey

The account registered as owner of this claim

bond_manager: Pubkey

The TicketManager this claim ticket was established under Determines the asset this ticket will be redeemed for

order_tag: OrderTag

The OrderTag associated with the creation of this struct

struck_timestamp: i64

The time slot during which the ticket was struck

maturation_timestamp: i64

The slot after which this claim can be redeemed for the underlying value

principal: u64

The total number of principal tokens the bond was struck for

interest: u64

The total number of interest tokens struck for this bond same underlying asset as the principal token

Implementations

impl SplitTicket

pub fn make_seeds<'a>(user: &'a [u8], bytes: &'a [u8]) -> [&'a [u8]; 3]

impl SplitTicket

pub fn verify_owner_manager(&self, owner: &Pubkey, manager: &Pubkey) -> Result<()>

Verify ticket ownership, takes owner and manager pubkeys

pub fn verify_owner(&self, owner: &Pubkey) -> Result<()>

Verify ticket ownership, takes owner pubkey